@@ -4,32 +4,31 @@ version: 0.0.1
 homepage: https://github.com/hpi-schul-cloud/schulcloud-flutter/tree/main/api
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  basics: ^0.4.0+1
-  collection: ^1.14.13
-  dartx: ^0.5.0
-  dio: ^3.0.10
-  enum_to_string: ^1.0.14
-  freezed_annotation: ^0.12.0
-  intl: ^0.16.1
-  json_annotation: ^3.1.1
-  meta: ^1.1.8
-  oxidized: ^3.1.0
-  rxdart: ^0.24.1
-  time_machine: ^0.9.13
+  collection: ^1.15.0
+  dartx: ^0.6.0
+  dio: ^4.0.0-beta7
+  enum_to_string: ^2.0.0-nullsafety.1
+  freezed_annotation: ^0.14.0
+  intl: ^0.17.0
+  json_annotation: ^4.0.0
+  meta: ^1.3.0
+  oxidized: ^4.0.0
+  rxdart: ^0.26.0
+  supercharged_dart: ^2.0.0
 
 dev_dependencies:
-  build_runner: ^1.10.2
-  extra_pedantic: ^1.2.0
-  freezed: ^0.12.6
-  json_serializable: ^3.5.1
-  test: ^1.6.0
+  build_runner: ^1.11.5
+  extra_pedantic: ^1.3.0
+  freezed: ^0.14.0
+  json_serializable: ^4.0.2
+  test: ^1.16.7
 
 dependency_overrides:
-  # https://github.com/Dana-Ferguson/time_machine/issues/44
-  time_machine:
-    git:
-      url: https://github.com/Dana-Ferguson/time_machine
-      ref: 7753e0d
+  # TODO(JonasWanke): remove these when build_runner updates its dependencies
+  http_multi_server: ^3.0.0
+  io: ^1.0.0
+  shelf_web_socket: ^1.0.0
+  web_socket_channel: ^2.0.0
