@@ -1,23 +1,23 @@
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:oxidized/oxidized.dart';
 
 import '../entity.dart';
-import '../errors.dart';
+import '../network.dart';
 import '../services/user.dart';
 import '../shallow.dart';
+import '../utils.dart';
 import 'data.dart';
 
 class ShallowAuthentication {
   ShallowAuthentication(Shallow shallow) : _shallow = shallow;
 
   final Shallow _shallow;
-  String _jwt;
+  String? _jwt;
 
-  Id<User> _currentUserId;
-  Id<User> get currentUserId => _currentUserId;
+  Id<User>? _currentUserId;
+  Id<User>? get currentUserId => _currentUserId;
   bool get isSignedIn => currentUserId != null;
 
   static const noAuthenticationHeader = 'x-no-authentication';
@@ -27,7 +27,7 @@ class ShallowAuthentication {
         if (options.headers.containsKey(noAuthenticationHeader)) {
           options.headers.remove(noAuthenticationHeader);
         } else {
-          options.headers['Authorization'] = 'Bearer $_jwt';
+          options.headers['Authorization'] = 'Bearer ${_jwt!}';
         }
         return options;
       },
@@ -36,41 +36,32 @@ class ShallowAuthentication {
 
   // sign-in
 
-  Future<Result<void, ShallowError>> signIn(AuthenticationBody body) async {
-    Response<dynamic> rawResponse;
-    try {
-      rawResponse = await _shallow.dio.post<dynamic>(
+  Future<Result<void, ShallowError>> signIn(AuthenticationBody body) {
+    final rawResponse = _shallow.dio.makeRequest<dynamic>(
+      (it) => it.post<dynamic>(
         '/authentication',
         data: body.toJson(),
         options: Options(
           headers: <String, dynamic>{noAuthenticationHeader: true},
         ),
-      );
-    } on DioError catch (e) {
-      print(e);
-      if (e.response.statusCode == HttpStatus.unauthorized) {
-        return Result.err(InvalidCredentialsError());
-      }
-      rethrow;
-    }
-    print(rawResponse);
-
-    final response = AuthenticationResponse.fromJson(
-        rawResponse.data as Map<String, dynamic>);
-    await signInWithJwt(response.accessToken);
-    return Result.ok(null);
+      ),
+    );
+    return rawResponse.fold(
+      (it) async {
+        final response =
+            AuthenticationResponse.fromJson(it.data! as Map<String, dynamic>);
+        await signInWithJwt(response.accessToken);
+      },
+      (it) => it is UnauthorizedError ? InvalidCredentialsError() : it,
+    );
   }
 
   Future<void> signInWithJwt(String jwt) async {
-    assert(jwt != null);
-
     _currentUserId = _decodeUserIdFromJwt(jwt);
     _jwt = jwt;
   }
 
   Id<User> _decodeUserIdFromJwt(String jwt) {
-    assert(jwt != null);
-
     final payloadString = String.fromCharCodes(base64Decode(jwt.split('.')[1]));
     final payloadJson = json.decode(payloadString) as Map<String, dynamic>;
     return Id<User>(payloadJson['userId'] as String);
@@ -79,20 +70,19 @@ class ShallowAuthentication {
   // sign-out
 
   Future<Result<void, ShallowError>> signOut() async {
+    final Result<void, ShallowError> result;
     try {
-      await _shallow.dio.delete<void>('/authentication');
-    } on DioError catch (e) {
-      if (e.response.statusCode == HttpStatus.unauthorized) {
-        // We were signed out already, e.g., because our JWT-token was no longer
-        // valid.
-      } else {
-        rethrow;
-      }
+      result = await _shallow.dio
+          .makeRequest<void>((it) => it.delete('/authentication'))
+          .map((it) => null)
+          .orElse(
+            (it) => it is UnauthorizedError ? Result.ok(null) : Result.err(it),
+          );
     } finally {
       _jwt = null;
       _currentUserId = null;
     }
-    return Result.ok(null);
+    return result;
   }
 }
 
