@@ -1,10 +1,8 @@
-import 'dart:io';
-
 import 'package:dio/dio.dart';
 import 'package:oxidized/oxidized.dart';
 
 import 'authentication/module.dart';
-import 'errors.dart';
+import 'network.dart';
 import 'services/course.dart';
 import 'services/lesson.dart';
 import 'services/me.dart';
@@ -14,10 +12,9 @@ import 'services/user.dart';
 
 class Shallow {
   Shallow({
-    Dio dio,
+    Dio? dio,
     this.apiRoot = 'https://api.hpi-schul-cloud.de',
-  })  : dio = dio ?? Dio(),
-        assert(apiRoot != null) {
+  }) : dio = dio ?? Dio() {
     this.dio.options.baseUrl = apiRoot;
     _authentication = ShallowAuthentication(this);
     this.dio.interceptors.add(authentication.dioInterceptor);
@@ -32,38 +29,27 @@ class Shallow {
   final Dio dio;
   final String apiRoot;
 
-  ShallowAuthentication _authentication;
+  late final ShallowAuthentication _authentication;
   ShallowAuthentication get authentication => _authentication;
 
-  CourseCollection _courses;
+  late final CourseCollection _courses;
   CourseCollection get courses => _courses;
 
   Future<Result<Me, ShallowError>> me() async {
-    Response<Map<String, dynamic>> response;
-    try {
-      response = await dio.get<Map<String, dynamic>>('/me');
-    } on DioError catch (e) {
-      switch (e.response.statusCode) {
-        case HttpStatus.unauthorized:
-          return Result.err(UnauthorizedError());
-        case HttpStatus.notFound:
-          return Result.err(NotFoundError());
-      }
-      rethrow;
-    }
-
-    return Result.ok(Me.fromJson(response.data));
+    final response =
+        await dio.makeRequest<Map<String, dynamic>>((it) => it.get('/me'));
+    return response.map((it) => Me.fromJson(it.data!));
   }
 
-  LessonCollection _lessons;
+  late final LessonCollection _lessons;
   LessonCollection get lessons => _lessons;
 
-  ArticleCollection _news;
+  late final ArticleCollection _news;
   ArticleCollection get news => _news;
 
-  UserCollection _users;
+  late final UserCollection _users;
   UserCollection get users => _users;
 
-  SchoolCollection _schools;
+  late final SchoolCollection _schools;
   SchoolCollection get schools => _schools;
 }
